chore: change how the http client is provided
