+++ composeApp/src/commonMain/kotlin/sources/remotesource/RemoteModule.kt
@@ -2,7 +2,7 @@ package sources.remotesource
 
 import io.ktor.client.HttpClient
 import org.koin.dsl.module
-import sources.remotesource.client.getHttpClient
+import sources.remotesource.client.createHttpClient
 import sources.remotesource.source.titles.TitlesRemoteSource
 import sources.remotesource.source.titles.TitlesRemoteSourceImpl
 
@@ -14,6 +14,6 @@ import sources.remotesource.source.titles.TitlesRemoteSourceImpl
  */
 
 val remoteModule = module {
-    single<HttpClient> { getHttpClient() }
+    single<HttpClient> { createHttpClient() }
     single<TitlesRemoteSource> { TitlesRemoteSourceImpl(client = get()) }
 }
