+++ composeApp/src/commonMain/kotlin/sources/remotesource/client/HttpClient.kt
@@ -1,6 +1,16 @@
 package sources.remotesource.client
 
 import io.ktor.client.HttpClient
+import io.ktor.client.plugins.DefaultRequest
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.client.plugins.logging.LogLevel
+import io.ktor.client.plugins.logging.Logger
+import io.ktor.client.plugins.logging.Logging
+import io.ktor.client.request.header
+import io.ktor.http.ContentType
+import io.ktor.http.HttpHeaders
+import io.ktor.serialization.kotlinx.json.json
+import kotlinx.serialization.json.Json
 
 /**
  * project : rhyme
@@ -9,4 +19,24 @@ import io.ktor.client.HttpClient
  * email   : mambobryan@gmail.com
  */
 
-expect fun getHttpClient(): HttpClient
+fun createHttpClient() : HttpClient = HttpClient {
+    install(ContentNegotiation) {
+        json(json = Json {
+            ignoreUnknownKeys = true
+        })
+    }
+
+    install(Logging) {
+        level = LogLevel.BODY
+        logger = object : Logger {
+            override fun log(message: String) {
+                println(message)
+            }
+        }
+    }
+
+    install(DefaultRequest) {
+        header(HttpHeaders.Accept, ContentType.Application.Json)
+        header(HttpHeaders.ContentType, ContentType.Application.Json)
+    }
+}
